requirements.txt
flask
python-dotenv
google-generativeai==0.5.0
